still working and adding few features

+++ database/migrations/2024_11_05_092434_create_wallets_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(User::class)->constrained();
             $table->string('currency')->default('NGN');
-            $table->bigInteger('balance')->unsigned()->default(10_000);
+            $table->float('balance')->unsigned()->default(10_000);
             $table->timestamps();
         });
     }
